+++ source/extensions/omni.flux.stage_manager.plugin.tree.usd/omni/flux/stage_manager/plugin/tree/usd/base/__init__.py
@@ -15,4 +15,9 @@
 * limitations under the License.
 """
 
-from .usd_base import StageManagerUSDTreePlugin
+from .usd_base import (
+    StageManagerUSDTreeDelegate,
+    StageManagerUSDTreeItem,
+    StageManagerUSDTreeModel,
+    StageManagerUSDTreePlugin,
+)