+++ source/extensions/omni.flux.stage_manager.core/docs/CHANGELOG.md
@@ -1,6 +1,13 @@
 # Changelog
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [1.1.1]
+### Changed
+- Set context instead of setting context items in the interaction plugins
+
+### Fixed
+- Fixed model serialization
+
 ## [1.1.0]
 ### Added
 - Added core setup function