+++ source/extensions/omni.flux.stage_manager.plugin.filter.usd/docs/CHANGELOG.md
@@ -1,6 +1,13 @@
 # Changelog
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [1.1.1]
+### Added
+- Added `context_name` field to the USD base
+
+### Fixed
+- Fixed dependencies
+
 ## [1.1.0]
 ### Added
 - Added `IgnorePrimsFilterPlugin` plugin