+++ source/extensions/omni.flux.stage_manager.plugin.widget.usd/docs/CHANGELOG.md
@@ -1,6 +1,14 @@
 # Changelog
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [1.2.0]
+### Changed
+- Added `context_name` field to the USD base
+- Implemented the `IsVisibleStateWidgetPlugin` plugin
+
+### Fixed
+- Fixed dependencies
+
 ## [1.1.0]
 ### Added
 - Added `StageManagerStateWidgetPlugin` plugin base